feat(fix some design issues): fix sidebar and some others design issues
