+++ src/utils/ulits.ts
@@ -3,7 +3,6 @@ export function openTabs(tabsName, className) {
   const x: any = document.getElementsByClassName("task-tab")
   const tabLabel: any = document.getElementsByClassName(className)
   console.log(tabLabel[0])
-  tabLabel[0].style.color = "#157BFF"
   for (i = 0; i < x.length; i++) {
     x[i].style.display = "none"
   }
